backend/requirements.txt
# C:\Nithin\AI-Projects\Neural-System-Decompiler\backend\requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
networkx==3.2.1
tree-sitter==0.22.3
tree-sitter-python==0.21.0
qdrant-client==1.8.0
fastembed==0.2.6
pymongo==4.6.2
langgraph==0.0.30
langchain-core==0.1.33
groq==0.4.2
tavily-python==0.3.3
pydantic==2.6.4